+++ sys/display_test.go
@@ -1,0 +1,12 @@
+package sys
+
+import "testing"
+
+func TestDisplay(t *testing.T) {
+	// TODO: DisplayOutput
+	// TODO: DisplayResult
+	// TODO: DisplayError
+	// TODO: DisplayData
+	// TODO: DisplayList
+	// TODO: displayOutput
+}