+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   * Conf package
   * conf.LoadEnv
   * dict.Inspect
-  * Sys package 
+  * Sys package and display functions
   * Root package
 ## v0.3.6 - Schema Combos 
   * Commit: 2026-04-01 22:44